extract options base class
