+++ superbench/benchmarks/micro_benchmarks/directx_gemm_flops_performance/Main.cpp
@@ -4,7 +4,8 @@
 #include "GPUCore.h"
 
 int main(int argc, char *argv[]) {
-    Options *opts = new Options(argc, argv);
-    auto gpucopy = new GPUCore(opts);
+    std::unique_ptr<GPUCoreOptions> opts = std::make_unique<GPUCoreOptions>(argc, argv);
+    opts->init();
+    std::unique_ptr<GPUCore> gpucopy = std::make_unique<GPUCore>(opts.get());
     gpucopy->Run();
 }
