+++ superbench/benchmarks/micro_benchmarks/directx_utils/Options.h
@@ -7,15 +7,8 @@
 #include <sstream>
 #include <string>
 
-namespace Option {
-enum Precision {
-    F16,
-    F32,
-};
-using PrecisionType = Option::Precision;
-} // namespace Option
-
-struct Options {
+class Options {
+  protected:
     char **begin;
     char **end;
 
@@ -43,7 +36,7 @@ struct Options {
             try {
                 return std::stoi(value);
             } catch (const std::exception &e) {
-                std::cerr << "Error: Invalid argument - " << option << " should be INT" << '\n';
+                std::cerr << "Error: Invalid argument - " << option << " should be INT " << e.what() << '\n';
                 exit(1);
             }
         }
@@ -81,59 +74,40 @@ struct Options {
      */
     bool cmd_option_exists(const std::string &option) { return std::find(begin, end, option) != end; }
 
-    void get_option_usage() {
-        std::cout << "Usage: " << std::endl;
-        std::cout << "  --help: Print help message." << std::endl;
-        std::cout << "  --num_loops: The number of benchmark runs." << std::endl;
-        std::cout << "  --num_warm_up: The number of warmup runs." << std::endl;
-        std::cout << "  --m: m dimension of GEMM." << std::endl;
-        std::cout << "  --n: n dimension of GEMM." << std::endl;
-        std::cout << "  --k: l dimension of GEMM." << std::endl;
-        std::cout << "  --fp16: half precision to compute." << std::endl;
-        std::cout << "  --fp32: float precision to compute." << std::endl;
-    }
+    /**
+     * @brief Get the option usage.
+     */
+    virtual void get_option_usage(){};
+
+    /**
+     * @brief Parse the arguments.
+     */
+    virtual void parse_arguments(){};
 
   public:
-    // Number of warm up rounds to run.
-    int num_warm_up = 0;
-    // The number of benchmark runs.
-    int num_loops = 0;
-    // Dimension m of GEMM.
-    int m = 0;
-    // Dimension n of GEMM.
-    int n = 0;
-    // Dimension k of GEMM.
-    int k = 0;
-    // The precision of calculate.
-    Option::PrecisionType mode_precision = Option::F32;
-
     /**
      * @brief Construct a new Command Line object.
      * @param argc the number of command line arguments.
      * @param argv the string array of comamnd line arguments.
      */
     Options(int argc, char *argv[]) {
+        begin = argv;
+        end = argv + argc;
+    }
+
+    /**
+     * @brief Init and parse the arguments.
+     */
+    virtual void init() {
+        if (cmd_option_exists("--help")) {
+            get_option_usage();
+            exit(0);
+        }
         try {
-            begin = argv;
-            end = argv + argc;
-            if (cmd_option_exists("--help")) {
-                get_option_usage();
-            } else {
-                num_loops = get_cmd_line_argument_int("--num_loops", 10);
-                num_warm_up = get_cmd_line_argument_int("--num_loops", 0);
-                m = get_cmd_line_argument_int("--m", 16 * 256);
-                n = get_cmd_line_argument_int("--n", 16 * 256);
-                k = get_cmd_line_argument_int("--k", 16 * 256);
-                if (get_cmd_line_argument_bool("--f16")) {
-                    mode_precision = Option::F16;
-                }
-                if (get_cmd_line_argument_bool("--f32")) {
-                    mode_precision = Option::F32;
-                }
-            }
+            parse_arguments();
         } catch (const std::exception &e) {
             std::cerr << "Error: Invalid argument - " << e.what() << '\n';
             exit(1);
         }
-    }
+    };
 };
